Allow custom handler chain.
